OTHER: Changed default skin to Almond-blue
I don't know under which license CleanAMP is provided, and I don't want to take
any risks, so I had to replace it. As the default skin is included in the
binary, it might be even necessary for it to be GPL, so I choose the debranded
Almond-blue skin for the time beeing.

Added creator of Almond-blue to AUTHORS file

Also added some magic to make it easier to add new skins through the resource
system